Updated buttons to be unselectable when not needed.
If no client is selected in the Server GUI APP, all the Action buttons will be greyed out and unselectable. If a client is selected, the 'Start Camera' button is shown if the camera is off, and the 'Shutdown Camera' button is shown if the camera is on. If the camera is on, all buttons are usable.